+++ src/components/Layout/Footer.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const Footer = () => {
+  return (
+    <div>
+       <footer>
+        <ul>
+            <li>Terms of Service</li>
+            <li>About Us</li>
+            <li>Download the app</li>
+            <li>Contact us</li>
+            <li>How it works</li>
+            <li>Privacy Policy</li>
+        </ul>
+      </footer>
+    </div>
+  );
+};
+
+export default Footer;
