+++ src/components/Layout/Header.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <div>
+      <h1>SquadMate</h1>
+    </div>
+  );
+};
+
+export default Header;
